+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added additional hint solver techniques
   - wings (xy-wing, xyz-wing, w-wing)
   - single digit patterns (empty rectangle, turbot fish)
+  - finned fish (finned x-wing / swordfish / jellyfish)
  
 ### Fixed